extract spec grammar regression into SpecGrammarKit
